Commit: 82
Successfully implemented the publish function (again), which uses the
ajax function to save changes in another way.  My intention is to use
this to implement a system where all data is saved upon input, but not
actually published to the actual live webpage until the user clicks the
"publish" button.